+++ runtime/onert/core/src/exec/SingleModelExecutors.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Samsung Electronics Co., Ltd. All Rights Reserved
+ * Copyright (c) 2023 Samsung Electronics Co., Ltd. All Rights Reserved
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,26 +14,11 @@
  * limitations under the License.
  */
 
-#ifndef __ONERT_EXEC_EXECUTORS_H__
-#define __ONERT_EXEC_EXECUTORS_H__
+#ifndef __ONERT_EXEC_SINGLE_MODEL_EXECUTORS_H__
+#define __ONERT_EXEC_SINGLE_MODEL_EXECUTORS_H__
 
 #include "exec/IExecutors.h"
 #include "ir/NNPkg.h"
-
-namespace std
-{
-
-template <> struct hash<std::pair<::onert::ir::ModelIndex, ::onert::ir::SubgraphIndex>>
-{
-  size_t
-  operator()(const std::pair<::onert::ir::ModelIndex, ::onert::ir::SubgraphIndex> &pair) const
-    noexcept
-  {
-    return (hash<uint32_t>()(pair.first.value()) << 16) ^ hash<uint32_t>()(pair.second.value());
-  }
-};
-
-} // namespace std
 
 namespace onert
 {
@@ -41,18 +26,24 @@ namespace exec
 {
 
 /**
- * @brief Class to gather executors
+ * @brief Class to gather executor set for single model NN package
  */
-class Executors : public IExecutors
+class SingleModelExecutors : public IExecutors
 {
 public:
-  Executors(void) = delete;
-  Executors(std::unique_ptr<ir::ModelEdges> model_edges) { _model_edges = std::move(model_edges); }
-  Executors(const Executors &) = delete;
-  Executors(Executors &&) = default;
-  ~Executors() = default;
+  /**
+   * @brief Construct a new SingleModelExecutors object
+   */
+  SingleModelExecutors(void) = default;
+  SingleModelExecutors(const SingleModelExecutors &) = delete;
+  SingleModelExecutors(SingleModelExecutors &&) = default;
 
-  // TODO Use Executor index
+  /**
+   * @brief Destroy the SingleModelExecutors object
+   */
+  ~SingleModelExecutors() = default;
+
+public:
   void emplace(const ir::ModelIndex &model_index, const ir::SubgraphIndex &subg_index,
                std::unique_ptr<IExecutor> exec) override;
 
@@ -70,17 +61,10 @@ public:
   void execute(const IODescription &desc) override;
 
 private:
-  void checkSupportedMultimodel() const;
-  uint16_t modelCount() const;
-
-private:
-  std::unordered_map<std::pair<ir::ModelIndex, ir::SubgraphIndex>, std::unique_ptr<IExecutor>>
-    _executors;
-  // NOTE _model_edges may use different struct type for executor implementation
-  std::unique_ptr<ir::ModelEdges> _model_edges;
+  std::unordered_map<ir::SubgraphIndex, std::unique_ptr<IExecutor>> _executors;
 };
 
 } // namespace exec
 } // namespace onert
 
-#endif // __ONERT_EXEC_EXECUTORS_H__
+#endif // __ONERT_EXEC_SINGLE_MODEL_EXECUTORS_H__
