+++ runtime/onert/core/src/interp/InterpExecutor.test.cc
@@ -15,7 +15,7 @@
  */
 
 #include "InterpExecutor.h"
-#include "../exec/Executors.h"
+#include "../exec/SingleModelExecutors.h"
 
 #include "exec/Execution.h"
 #include "ir/Graph.h"
